requirements.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.1.0
Markdown==3.3.3
numpy==1.19.4
pandas==1.1.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
Serializer==0.2.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
zipp==3.4.0
